my-app/backend/requirements.txt
langchain==0.0.184
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
openai==0.27.6
# faiss-cpu==1.7.4
altair==4
# tiktoken==0.4.0
Flask
supabase
sentence-transformers
torch
transformers
pdfplumber
python-dotenv
tokenizers
flask_cors
sqlalchemy
psycopg2-binary
PyPDF2
requests
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client 
requests
PyMuPDF
scikit-learn